stdlib: Fix clang strtod build on 32 bits
clang 18+ issues the warning:

strtod_l.c:876:18: error: result of comparison of constant 9223372036854775807 with expression of type 'size_t' (aka 'unsigned int') is always true [-Werror,-Wtautological-constant-out-of-range-compare]
  876 |   assert (dig_no <= (uintmax_t) INTMAX_MAX);
      |   ~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~
../assert/assert.h:155:11: note: expanded from macro 'assert'
  155 |       if (__VA_ARGS__)                                                  \
      |           ^~~~~~~~~~~

Use a cast to expected type to avoid the warning.

+++ stdlib/strtod_l.c
@@ -873,7 +873,7 @@ ____STRTOF_INTERNAL (const STRING_TYPE *nptr, STRING_TYPE **endptr, int group,
 	  c = *++cp;
 	}
     }
-  assert (dig_no <= (uintmax_t) INTMAX_MAX);
+  assert ((uintmax_t)dig_no <= (uintmax_t) INTMAX_MAX);
 
   /* Remember start of exponent (if any).  */
   expp = cp;
@@ -903,7 +903,7 @@ ____STRTOF_INTERNAL (const STRING_TYPE *nptr, STRING_TYPE **endptr, int group,
 	    {
 	      if (exp_negative)
 		{
-		  assert (int_no <= (uintmax_t) (INTMAX_MAX
+		  assert ((uintmax_t)int_no <= (uintmax_t) (INTMAX_MAX
 						 + MIN_EXP - MANT_DIG) / 4);
 		  exp_limit = -MIN_EXP + MANT_DIG + 4 * (intmax_t) int_no;
 		}
@@ -912,7 +912,8 @@ ____STRTOF_INTERNAL (const STRING_TYPE *nptr, STRING_TYPE **endptr, int group,
 		  if (int_no)
 		    {
 		      assert (lead_zero == 0
-			      && int_no <= (uintmax_t) INTMAX_MAX / 4);
+			      && (uintmax_t)int_no <= (uintmax_t) INTMAX_MAX
+						      / 4);
 		      exp_limit = MAX_EXP - 4 * (intmax_t) int_no + 3;
 		    }
 		  else if (lead_zero == (size_t) -1)
@@ -923,7 +924,7 @@ ____STRTOF_INTERNAL (const STRING_TYPE *nptr, STRING_TYPE **endptr, int group,
 		    }
 		  else
 		    {
-		      assert (lead_zero
+		      assert ((uintmax_t)lead_zero
 			      <= (uintmax_t) (INTMAX_MAX - MAX_EXP - 3) / 4);
 		      exp_limit = (MAX_EXP
 				   + 4 * (intmax_t) lead_zero
@@ -935,7 +936,7 @@ ____STRTOF_INTERNAL (const STRING_TYPE *nptr, STRING_TYPE **endptr, int group,
 	    {
 	      if (exp_negative)
 		{
-		  assert (int_no
+		  assert ((uintmax_t)int_no
 			  <= (uintmax_t) (INTMAX_MAX + MIN_10_EXP - MANT_DIG));
 		  exp_limit = -MIN_10_EXP + MANT_DIG + (intmax_t) int_no;
 		}
@@ -944,7 +945,7 @@ ____STRTOF_INTERNAL (const STRING_TYPE *nptr, STRING_TYPE **endptr, int group,
 		  if (int_no)
 		    {
 		      assert (lead_zero == 0
-			      && int_no <= (uintmax_t) INTMAX_MAX);
+			      && (uintmax_t)int_no <= (uintmax_t) INTMAX_MAX);
 		      exp_limit = MAX_10_EXP - (intmax_t) int_no + 1;
 		    }
 		  else if (lead_zero == (size_t) -1)
@@ -955,7 +956,7 @@ ____STRTOF_INTERNAL (const STRING_TYPE *nptr, STRING_TYPE **endptr, int group,
 		    }
 		  else
 		    {
-		      assert (lead_zero
+		      assert ((uintmax_t)lead_zero
 			      <= (uintmax_t) (INTMAX_MAX - MAX_10_EXP - 1));
 		      exp_limit = MAX_10_EXP + (intmax_t) lead_zero + 1;
 		    }
